Update version and Battle System
I've updated the version numbers for both the actual plugin and readme
to match.

Battle system has been updated, still not 100% stable, but HP and XP
seems to be working good, and results page is created.

No current way to replenish HP which makes attacking useless once
players are dead. Will have to look into a Gold system, and perhaps
create a pluggable way to integrate it with Comments, Posts, or other
plugins.  Also wp_cron or a normal cron system needs to be integrated.