Updated the shop fix the display issues
Updated the shop to accurately show the items in each shop that's
available.
Each shop can be created with a specific name and description and have a
different variety of items that it can sell. This will later pave way
for a City module to allow you to place a Shop or variety of TYPES of
shops in each city to offer different thigns.